DONE: recordBoatReturn API GraphQL TODO: consider penalties on that API

+++ server/graphql/resolvers/advertisement.js
@@ -2,6 +2,7 @@ const Boat = require("../../models/boat")
 const {authenticated, authorization} = require("../../auth/auth");
 const {transformBoat} = require("./merge");
 const Rental = require("../../models/rental");
+const {boatNotFound} = require("../../helpers/problemMessages")
 
 module.exports = {
     advertisementsByShipowner: authenticated(authorization('shipowner')(async (_, {req}) => {
@@ -16,19 +17,25 @@ module.exports = {
             return boats.map(transformBoat)
         } catch (err) {throw new Error(`Can't find shipowner advertisements. ${err}`)}
     })),
-    withdrawAdvertisement: authenticated(authorization('shipowner')(async ({boatId}) => {
+    withdrawAdvertisement: authenticated(authorization('shipowner')(async ({boatId}, {req}) => {
         try {
-            await Boat.updateOne(
-                {_id: boatId},
-                {$unset: {'advertisement': 1} }
+            const result = await Boat.updateOne(
+                {
+                    $and: [
+                        {_id: boatId},
+                        {shipowner: req.userId}
+                    ]
+                },
+                {$unset: {'advertisement': 1}}
             )
+            if (result.deletedCount === 0) return {withdrawAdvertisementProblem: boatNotFound}
             await Rental.deleteMany({
                 $and: [
                     {boat: boatId},
-                    {fromDate: { $gt: new Date() } }
+                    {fromDate: {$gt: new Date()}}
                 ]
             })
-            return {withdrawnAdvertisementId: boatId }
+            return {withdrawnAdvertisementId: boatId}
         } catch (err) {throw new Error(`Can't withdraw advertisement. ${err}`)}
     })),
 }
